♻️(backend) rename `marsha.core.tests.utils`
Rename this module to be allowed to move the
`marsha.core.utils` module test into
a `marsha.core.tests.utils` submodule.
