+++ src/backend/marsha/.cookiecutter/{{cookiecutter.app_name}}/tests/test_api.py
@@ -9,7 +9,7 @@ from marsha.core.simple_jwt.factories import (
     PlaylistLtiTokenFactory,
     StudentLtiTokenFactory,
 )
-from marsha.core.tests.utils import reload_urlconf
+from marsha.core.tests.testing_utils import reload_urlconf
 
 from ..factories import {{cookiecutter.model}}Factory
 from ..models import {{cookiecutter.model}}
