+++ src/backend/marsha/bbb/tests/api/classroom/test_update.py
@@ -23,7 +23,7 @@ from marsha.core.simple_jwt.factories import (
     UserAccessTokenFactory,
 )
 from marsha.core.simple_jwt.tokens import ResourceAccessToken
-from marsha.core.tests.utils import reload_urlconf
+from marsha.core.tests.testing_utils import reload_urlconf
 
 
 # We don't enforce arguments documentation in tests
