+++ src/backend/marsha/bbb/tests/api/classroomdocument/test_initiate_upload.py
@@ -18,7 +18,7 @@ from marsha.core.simple_jwt.factories import (
     InstructorOrAdminLtiTokenFactory,
     UserAccessTokenFactory,
 )
-from marsha.core.tests.utils import reload_urlconf
+from marsha.core.tests.testing_utils import reload_urlconf
 
 
 # We don't enforce arguments documentation in tests
