+++ src/backend/marsha/bbb/tests/api/classroomdocument/test_options.py
@@ -7,7 +7,7 @@ from marsha.core.simple_jwt.factories import (
     StudentLtiTokenFactory,
     UserAccessTokenFactory,
 )
-from marsha.core.tests.utils import reload_urlconf
+from marsha.core.tests.testing_utils import reload_urlconf
 
 
 @override_settings(BBB_ENABLED=True)
