+++ src/backend/marsha/bbb/tests/api/classroomdocument/test_update.py
@@ -16,7 +16,7 @@ from marsha.core.simple_jwt.factories import (
     StudentLtiTokenFactory,
     UserAccessTokenFactory,
 )
-from marsha.core.tests.utils import reload_urlconf
+from marsha.core.tests.testing_utils import reload_urlconf
 
 
 # We don't enforce arguments documentation in tests
