+++ src/backend/marsha/bbb/tests/test_views_lti_select.py
@@ -8,7 +8,7 @@ from django.test import TestCase, override_settings
 
 from marsha.bbb.factories import ClassroomFactory
 from marsha.core.factories import PlaylistFactory
-from marsha.core.tests.utils import generate_passport_and_signed_lti_parameters
+from marsha.core.tests.testing_utils import generate_passport_and_signed_lti_parameters
 from marsha.core.utils.lti_select_utils import get_lti_select_resources
