+++ src/backend/marsha/core/tests/api/video/test_retrieve.py
@@ -23,7 +23,7 @@ from marsha.core.simple_jwt.factories import (
     StudentLtiTokenFactory,
     UserAccessTokenFactory,
 )
-from marsha.core.tests.utils import RSA_KEY_MOCK
+from marsha.core.tests.testing_utils import RSA_KEY_MOCK
 
 
 # This test module is quite long...
