+++ src/backend/marsha/core/tests/lti/tests.py
@@ -14,7 +14,7 @@ from marsha.core.factories import (
 from marsha.core.lti import LTI
 from marsha.core.lti.utils import get_or_create_resource
 from marsha.core.models import Video
-from marsha.core.tests.utils import generate_passport_and_signed_lti_parameters
+from marsha.core.tests.testing_utils import generate_passport_and_signed_lti_parameters
 
 
 # We don't enforce arguments documentation in tests
