+++ src/backend/marsha/core/tests/simple_jwt/test_tokens.py
@@ -14,7 +14,7 @@ from marsha.core.simple_jwt.tokens import (
     ResourceRefreshToken,
     UserAccessToken,
 )
-from marsha.core.tests.utils import generate_passport_and_signed_lti_parameters
+from marsha.core.tests.testing_utils import generate_passport_and_signed_lti_parameters
 
 
 class LTISelectFormAccessTokenTestCase(TestCase):
