+++ src/backend/marsha/core/tests/views/test_lti_respond.py
@@ -5,7 +5,7 @@ from unittest import mock
 from django.test import TestCase
 
 from marsha.core.simple_jwt.tokens import LTISelectFormAccessToken
-from marsha.core.tests.utils import generate_passport_and_signed_lti_parameters
+from marsha.core.tests.testing_utils import generate_passport_and_signed_lti_parameters
 
 
 # We don't enforce arguments documentation in tests
