+++ src/backend/marsha/deposit/tests/api/filedepositories/test_depositedfiles.py
@@ -16,7 +16,7 @@ from marsha.core.simple_jwt.factories import (
     StudentLtiTokenFactory,
     UserAccessTokenFactory,
 )
-from marsha.core.tests.utils import RSA_KEY_MOCK, reload_urlconf
+from marsha.core.tests.testing_utils import RSA_KEY_MOCK, reload_urlconf
 from marsha.deposit.factories import DepositedFileFactory, FileDepositoryFactory
