+++ src/backend/marsha/deposit/tests/api/filedepositories/test_lti_select.py
@@ -3,7 +3,7 @@ from django.test import TestCase, override_settings
 
 from marsha.core import factories as core_factories
 from marsha.core.simple_jwt.factories import PlaylistLtiTokenFactory
-from marsha.core.tests.utils import reload_urlconf
+from marsha.core.tests.testing_utils import reload_urlconf
 from marsha.deposit.factories import FileDepositoryFactory
