+++ src/backend/marsha/e2e/test_lti_markdown.py
@@ -10,7 +10,7 @@ import pytest
 from pytest_django.live_server_helper import LiveServer
 from waffle.testutils import override_switch
 
-from marsha.core.tests.utils import generate_passport_and_signed_lti_parameters
+from marsha.core.tests.testing_utils import generate_passport_and_signed_lti_parameters
 
 
 @pytest.mark.django_db()
